+++ scanapi/scan.py
@@ -1,14 +1,19 @@
 import logging
+import sys
 import yaml
 
 from scanapi.config_loader import load_config_file
 from scanapi.errors import (
+    BadConfigurationError,
     EmptyConfigFileError,
     FileFormatNotSupportedError,
     InvalidKeyError,
+    InvalidPythonCodeError,
     MissingMandatoryKeyError,
 )
+from scanapi.exit_code import ExitCode
 from scanapi.reporter import Reporter
+from scanapi.session import session
 from scanapi.settings import settings
 from scanapi.tree import EndpointNode
 from scanapi.tree.tree_keys import API_KEY, ROOT_SCOPE
@@ -24,14 +29,14 @@ def scan():
     except FileNotFoundError as e:
         error_message = f"Could not find API spec file: {spec_path}. {str(e)}"
         logger.error(error_message)
-        return
+        raise SystemExit(ExitCode.USAGE_ERROR)
     except EmptyConfigFileError as e:
         error_message = f"API spec file is empty. {str(e)}"
         logger.error(error_message)
-        return
+        raise SystemExit(ExitCode.USAGE_ERROR)
     except (yaml.YAMLError, FileFormatNotSupportedError) as e:
         logger.error(e)
-        return
+        raise SystemExit(ExitCode.USAGE_ERROR)
 
     try:
         if API_KEY not in api_spec:
@@ -40,13 +45,24 @@ def scan():
         root_node = EndpointNode(api_spec[API_KEY])
         results = root_node.run()
 
-    except (InvalidKeyError, MissingMandatoryKeyError, KeyError) as e:
+    except (
+        InvalidKeyError,
+        MissingMandatoryKeyError,
+        KeyError,
+        InvalidPythonCodeError,
+    ) as e:
         error_message = "Error loading API spec."
         error_message = "{} {}".format(error_message, str(e))
         logger.error(error_message)
-        return
+        raise SystemExit(ExitCode.USAGE_ERROR)
 
-    write_report(results)
+    try:
+        write_report(results)
+    except (BadConfigurationError, InvalidPythonCodeError) as e:
+        logger.error(e)
+        raise SystemExit(ExitCode.USAGE_ERROR)
+
+    session.exit()
 
 
 def write_report(results):