+++ scanapi/test_status.py
@@ -1,0 +1,12 @@
+class TestStatus:
+    __test__ = False
+    """
+    Encodes the valid test status.
+    """
+
+    #: test passed
+    PASSED = "passed"
+    #: test failed
+    FAILED = "failed"
+    #: test error
+    ERROR = "error"